+++ printer_test.go
@@ -35,6 +35,51 @@ func ExamplePrinter() {
 	s := p.Format(v)
 	fmt.Println(s)
 
+	// output: github.com/dogmatiq/dapper_test.TreeNode{
+	//     Name:     "root"
+	//     Value:    nil
+	//     Children: {
+	//         {
+	//             Name:     "branch #1"
+	//             Value:    int(100)
+	//             Children: nil
+	//         }
+	//         {
+	//             Name:     "branch #2"
+	//             Value:    github.com/dogmatiq/dapper_test.NodeValue{}
+	//             Children: nil
+	//         }
+	//     }
+	// }
+}
+
+func ExamplePrinter_Config() {
+	type TreeNode struct {
+		Name     string
+		Value    interface{}
+		Children []*TreeNode
+	}
+
+	type NodeValue struct{}
+
+	v := TreeNode{
+		Name: "root",
+		Children: []*TreeNode{
+			{
+				Name:  "branch #1",
+				Value: 100,
+			},
+			{
+				Name:  "branch #2",
+				Value: NodeValue{},
+			},
+		},
+	}
+
+	p := Printer{Config: Config{OmitPackagePaths: true}}
+	s := p.Format(v)
+	fmt.Println(s)
+
 	// output: dapper_test.TreeNode{
 	//     Name:     "root"
 	//     Value:    nil